added history read api
